Refactor First and Join seed node processes to shared logic for dry and logic hardening
