+++ akka-cluster/src/main/scala/akka/cluster/ClusterDaemon.scala
@@ -4,6 +4,12 @@
 
 package akka.cluster
 
+import scala.collection.immutable
+import scala.concurrent.duration._
+import scala.concurrent.Future
+import scala.concurrent.Promise
+import scala.util.control.NonFatal
+
 import akka.actor._
 import akka.annotation.InternalApi
 import akka.actor.SupervisorStrategy.Stop
@@ -15,16 +21,10 @@ import akka.pattern.ask
 import akka.remote.{ QuarantinedEvent => ClassicQuarantinedEvent }
 import akka.remote.artery.QuarantinedEvent
 import akka.util.Timeout
-import com.typesafe.config.Config
-import scala.collection.immutable
-import scala.concurrent.duration._
-import scala.concurrent.Future
-import scala.concurrent.Promise
-import scala.util.control.NonFatal
-
 import akka.event.ActorWithLogClass
 import akka.event.Logging
 import com.github.ghik.silencer.silent
+import com.typesafe.config.Config
 
 /**
  * Base trait for all cluster messages. All ClusterMessage's are serializable.
@@ -1536,275 +1536,6 @@ private[cluster] class ClusterCoreDaemon(publisher: ActorRef, joinConfigCompatCh
     publisher ! CurrentInternalStats(gossipStats, vclockStats)
   }
 
-}
-
-/**
- * INTERNAL API.
- *
- * Used only for the first seed node.
- * Sends InitJoin to all seed nodes (except itself).
- * If other seed nodes are not part of the cluster yet they will reply with
- * InitJoinNack or not respond at all and then the first seed node
- * will join itself to initialize the new cluster. When the first
- * seed node is restarted, and some other seed node is part of the cluster
- * it will reply with InitJoinAck and then the first seed node will join
- * that other seed node to join existing cluster.
- */
-@InternalApi
-private[cluster] final class FirstSeedNodeProcess(
-    seedNodes: immutable.IndexedSeq[Address],
-    joinConfigCompatChecker: JoinConfigCompatChecker)
-    extends Actor {
-  import InternalClusterAction._
-  import ClusterUserAction.JoinTo
-
-  val cluster = Cluster(context.system)
-  import cluster.settings._
-  import cluster.ClusterLogger._
-
-  def selfAddress = cluster.selfAddress
-
-  if (seedNodes.size <= 1 || seedNodes.head != selfAddress)
-    throw new IllegalArgumentException("Join seed node should not be done")
-
-  val timeout = Deadline.now + cluster.settings.SeedNodeTimeout
-
-  var remainingSeedNodes = seedNodes.toSet - selfAddress
-
-  // retry until one ack, or all nack, or timeout
-  import context.dispatcher
-  val retryTask = cluster.scheduler.scheduleWithFixedDelay(1.second, 1.second, self, JoinSeedNode)
-  self ! JoinSeedNode
-
-  override def postStop(): Unit = retryTask.cancel()
-
-  def receive = {
-    case JoinSeedNode =>
-      if (timeout.hasTimeLeft) {
-        val requiredNonSensitiveKeys =
-          JoinConfigCompatChecker.removeSensitiveKeys(joinConfigCompatChecker.requiredKeys, cluster.settings)
-        // configToValidate only contains the keys that are required according to JoinConfigCompatChecker on this node
-        val configToValidate =
-          JoinConfigCompatChecker.filterWithKeys(requiredNonSensitiveKeys, context.system.settings.config)
-        // send InitJoin to remaining seed nodes (except myself)
-        remainingSeedNodes.foreach { a =>
-          context.actorSelection(context.parent.path.toStringWithAddress(a)) ! InitJoin(configToValidate)
-        }
-      } else {
-        // no InitJoinAck received, initialize new cluster by joining myself
-        if (isDebugEnabled)
-          logDebug("Couldn't join other seed nodes, will join myself. seed-nodes=[{}]", seedNodes.mkString(", "))
-        context.parent ! JoinTo(selfAddress)
-        context.stop(self)
-      }
-
-    case InitJoinAck(address, CompatibleConfig(clusterConfig)) =>
-      logInfo("Received InitJoinAck message from [{}] to [{}]", sender(), selfAddress)
-      // validates config coming from cluster against this node config
-      joinConfigCompatChecker.check(clusterConfig, context.system.settings.config) match {
-        case Valid =>
-          // first InitJoinAck reply
-          context.parent ! JoinTo(address)
-          context.stop(self)
-
-        case Invalid(messages) if ByPassConfigCompatCheck =>
-          logWarning(
-            "Cluster validated this node config, but sent back incompatible settings: {}. " +
-            "Join will be performed because compatibility check is configured to not be enforced.",
-            messages.mkString(", "))
-          context.parent ! JoinTo(address)
-          context.stop(self)
-
-        case Invalid(messages) =>
-          logError(
-            "Cluster validated this node config, but sent back incompatible settings: {}. " +
-            "It's recommended to perform a full cluster shutdown in order to deploy this new version. " +
-            "If a cluster shutdown isn't an option, you may want to disable this protection by setting " +
-            "'akka.cluster.configuration-compatibility-check.enforce-on-join = off'. " +
-            "Note that disabling it will allow the formation of a cluster with nodes having incompatible configuration settings. " +
-            "This node will be shutdown!",
-            messages.mkString(", "))
-          context.stop(self)
-          CoordinatedShutdown(context.system).run(CoordinatedShutdown.IncompatibleConfigurationDetectedReason)
-      }
-
-    case InitJoinAck(address, UncheckedConfig) =>
-      logInfo("Received InitJoinAck message from [{}] to [{}]", sender(), selfAddress)
-      logWarning("Joining a cluster without configuration compatibility check feature.")
-      context.parent ! JoinTo(address)
-      context.stop(self)
-
-    case InitJoinAck(address, IncompatibleConfig) =>
-      // first InitJoinAck reply, but incompatible
-      if (ByPassConfigCompatCheck) {
-        // only join if set to ignore config validation
-        logInfo("Received InitJoinAck message from [{}] to [{}]", sender(), selfAddress)
-        logWarning(
-          "Joining cluster with incompatible configurations. " +
-          "Join will be performed because compatibility check is configured to not be enforced.")
-        context.parent ! JoinTo(address)
-        context.stop(self)
-      } else {
-        logError(
-          "Couldn't join seed nodes because of incompatible cluster configuration. " +
-          "It's recommended to perform a full cluster shutdown in order to deploy this new version." +
-          "If a cluster shutdown isn't an option, you may want to disable this protection by setting " +
-          "'akka.cluster.configuration-compatibility-check.enforce-on-join = off'. " +
-          "Note that disabling it will allow the formation of a cluster with nodes having incompatible configuration settings. " +
-          "This node will be shutdown!")
-        context.stop(self)
-        CoordinatedShutdown(context.system).run(CoordinatedShutdown.IncompatibleConfigurationDetectedReason)
-      }
-
-    case InitJoinNack(address) =>
-      logInfo("Received InitJoinNack message from [{}] to [{}]", sender(), selfAddress)
-      remainingSeedNodes -= address
-      if (remainingSeedNodes.isEmpty) {
-        // initialize new cluster by joining myself when nacks from all other seed nodes
-        context.parent ! JoinTo(selfAddress)
-        context.stop(self)
-      }
-  }
-
-}
-
-/**
- * INTERNAL API.
- *
- * Sends InitJoin to all seed nodes (except itself) and expect
- * InitJoinAck reply back. The seed node that replied first
- * will be used, joined to. InitJoinAck replies received after the
- * first one are ignored.
- *
- * Retries if no InitJoinAck replies are received within the
- * SeedNodeTimeout.
- * When at least one reply has been received it stops itself after
- * an idle SeedNodeTimeout.
- *
- * The seed nodes can be started in any order, but they will not be "active",
- * until they have been able to join another seed node (seed1).
- * They will retry the join procedure.
- * So one possible startup scenario is:
- * 1. seed2 started, but doesn't get any ack from seed1 or seed3
- * 2. seed3 started, doesn't get any ack from seed1 or seed3 (seed2 doesn't reply)
- * 3. seed1 is started and joins itself
- * 4. seed2 retries the join procedure and gets an ack from seed1, and then joins to seed1
- * 5. seed3 retries the join procedure and gets acks from seed2 first, and then joins to seed2
- *
- */
-@InternalApi
-private[cluster] final class JoinSeedNodeProcess(
-    seedNodes: immutable.IndexedSeq[Address],
-    joinConfigCompatChecker: JoinConfigCompatChecker)
-    extends Actor {
-  import InternalClusterAction._
-  import ClusterUserAction.JoinTo
-
-  val cluster = Cluster(context.system)
-  import cluster.settings._
-  import cluster.ClusterLogger._
-  def selfAddress = cluster.selfAddress
-
-  if (seedNodes.isEmpty || seedNodes.head == selfAddress)
-    throw new IllegalArgumentException("Join seed node should not be done")
-
-  context.setReceiveTimeout(SeedNodeTimeout)
-
-  var attempt = 0
-
-  // all seed nodes, except this one
-  val otherSeedNodes = seedNodes.toSet - selfAddress
-
-  override def preStart(): Unit = self ! JoinSeedNode
-
-  def receive = {
-    case JoinSeedNode =>
-      val requiredNonSensitiveKeys =
-        JoinConfigCompatChecker.removeSensitiveKeys(joinConfigCompatChecker.requiredKeys, cluster.settings)
-      // configToValidate only contains the keys that are required according to JoinConfigCompatChecker on this node
-      val configToValidate =
-        JoinConfigCompatChecker.filterWithKeys(requiredNonSensitiveKeys, context.system.settings.config)
-      // send InitJoin to all seed nodes (except myself)
-      attempt += 1
-      otherSeedNodes.foreach { a =>
-        context.actorSelection(context.parent.path.toStringWithAddress(a)) ! InitJoin(configToValidate)
-      }
-
-    case InitJoinAck(address, CompatibleConfig(clusterConfig)) =>
-      logInfo("Received InitJoinAck message from [{}] to [{}]", sender(), selfAddress)
-      // validates config coming from cluster against this node config
-      joinConfigCompatChecker.check(clusterConfig, context.system.settings.config) match {
-        case Valid =>
-          // first InitJoinAck reply
-          context.parent ! JoinTo(address)
-          context.become(done)
-
-        case Invalid(messages) if ByPassConfigCompatCheck =>
-          logWarning(
-            "Cluster validated this node config, but sent back incompatible settings: {}. " +
-            "Join will be performed because compatibility check is configured to not be enforced.",
-            messages.mkString(", "))
-          context.parent ! JoinTo(address)
-          context.become(done)
-
-        case Invalid(messages) =>
-          logError(
-            "Cluster validated this node config, but sent back incompatible settings: {}. " +
-            "It's recommended to perform a full cluster shutdown in order to deploy this new version. " +
-            "If a cluster shutdown isn't an option, you may want to disable this protection by setting " +
-            "'akka.cluster.configuration-compatibility-check.enforce-on-join = off'. " +
-            "Note that disabling it will allow the formation of a cluster with nodes having incompatible configuration settings. " +
-            "This node will be shutdown!",
-            messages.mkString(", "))
-          context.stop(self)
-          CoordinatedShutdown(context.system).run(CoordinatedShutdown.IncompatibleConfigurationDetectedReason)
-      }
-
-    case InitJoinAck(address, UncheckedConfig) =>
-      logWarning("Joining a cluster without configuration compatibility check feature.")
-      context.parent ! JoinTo(address)
-      context.become(done)
-
-    case InitJoinAck(address, IncompatibleConfig) =>
-      // first InitJoinAck reply, but incompatible
-      if (ByPassConfigCompatCheck) {
-        logInfo("Received InitJoinAck message from [{}] to [{}]", sender(), selfAddress)
-        logWarning(
-          "Joining cluster with incompatible configurations. " +
-          "Join will be performed because compatibility check is configured to not be enforced.")
-        // only join if set to ignore config validation
-        context.parent ! JoinTo(address)
-        context.become(done)
-      } else {
-        logError(
-          ClusterLogMarker.joinFailed,
-          "Couldn't join seed nodes because of incompatible cluster configuration. " +
-          "It's recommended to perform a full cluster shutdown in order to deploy this new version." +
-          "If a cluster shutdown isn't an option, you may want to disable this protection by setting " +
-          "'akka.cluster.configuration-compatibility-check.enforce-on-join = off'. " +
-          "Note that disabling it will allow the formation of a cluster with nodes having incompatible configuration settings. " +
-          "This node will be shutdown!")
-        context.stop(self)
-        CoordinatedShutdown(context.system).run(CoordinatedShutdown.IncompatibleConfigurationDetectedReason)
-      }
-
-    case InitJoinNack(_) => // that seed was uninitialized
-
-    case ReceiveTimeout =>
-      if (attempt >= 2)
-        logWarning(
-          ClusterLogMarker.joinFailed,
-          "Couldn't join seed nodes after [{}] attempts, will try again. seed-nodes=[{}]",
-          attempt,
-          seedNodes.filterNot(_ == selfAddress).mkString(", "))
-      // no InitJoinAck received, try again
-      self ! JoinSeedNode
-  }
-
-  def done: Actor.Receive = {
-    case InitJoinAck(_, _) => // already received one, skip rest
-    case ReceiveTimeout    => context.stop(self)
-  }
 }
 
 /**
